+++ command.js
@@ -1,10 +1,17 @@
 const spec = require('conventional-changelog-config-spec')
 const { getConfiguration } = require('./lib/configuration')
 const defaults = require('./defaults')
-const { START_OF_LAST_RELEASE_PATTERN } = require('./lib/lifecycles/changelog')
 
 const yargs = require('yargs')
   .usage('Usage: $0 [options]')
+  .option('packageFiles', {
+    default: defaults.packageFiles,
+    array: true
+  })
+  .option('bumpFiles', {
+    default: defaults.bumpFiles,
+    array: true
+  })
   .option('release-as', {
     alias: 'r',
     describe: 'Specify the release type manually (like npm version <major|minor|patch>)',
@@ -77,7 +84,7 @@ const yargs = require('yargs')
   .option('git-tag-fallback', {
     type: 'boolean',
     default: defaults.gitTagFallback,
-    describe: `fallback to git tags for version, if no meta-information file is found (e.g., package.json)`
+    describe: 'fallback to git tags for version, if no meta-information file is found (e.g., package.json)'
   })
   .option('path', {
     type: 'string',
@@ -85,7 +92,7 @@ const yargs = require('yargs')
   })
   .option('changelogHeader', {
     type: 'string',
-    describe: 'Use a custom header when generating and updating changelog.'
+    describe: '[DEPRECATED] Use a custom header when generating and updating changelog.\nThis option will be removed in the next major version, please use --header.'
   })
   .option('preset', {
     type: 'string',
@@ -108,20 +115,13 @@ const yargs = require('yargs')
   .pkgConf('standard-version')
   .config(getConfiguration())
   .wrap(97)
-  .check((args) => {
-    if (args.changelogHeader && args.changelogHeader.search(START_OF_LAST_RELEASE_PATTERN) !== -1) {
-      throw Error(`custom changelog header must not match ${START_OF_LAST_RELEASE_PATTERN}`)
-    } else {
-      return true
-    }
-  })
 
 Object.keys(spec.properties).forEach(propertyKey => {
   const property = spec.properties[propertyKey]
   yargs.option(propertyKey, {
     type: property.type,
     describe: property.description,
-    default: property.default,
+    default: defaults[propertyKey] ? defaults[propertyKey] : property.default,
     group: 'Preset Configuration:'
   })
 })